Use bgr8 encoding for loading images
OpenCV assumes every image is stored in BGR8 format.
When we write an image to disk using cv2.imwrite,
it will also assume this format.

However PIL reads images as RGB8.

In order to make both OpenCV and PIL compatible,
we need to read the image as bgr8 so that when
it writes to disk all the channels are correctly
ordered to be read by PIL.